pyproject.toml
[build-system]
requires = ["setuptools>=61.2", "setuptools_scm>=8", "wheel"]
build-backend = "setuptools.build_meta"
[tool.setuptools_scm]
# For smarter version schemes and other configuration options,
# check out https://github.com/pypa/setuptools_scm
version_scheme = "no-guess-dev"

[project]
name = "imputegaps"
description = "Impute data gaps in a pandas dataframe"
authors = [{name = "EMSK", email = "EMSK@cbs.nl"}, {"name"= "EVLT", "email"= "evlt@cbs.nl"}]
dynamic = ["version"]
requires-python = ">=3.10"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
]
dependencies = [
    "pandas",
    "pyyaml",
]

# install tox manually into your venv or on your system
# in case you use uv, install with
# uv tool install tox --with tox-uv
# in which case venvs are build using uv
[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-compare[pandas]",
    "black",
    "pre-commit",
    "pylint",
    "ruff>=0.9.3",
    "isort>=5.13.2",
]

[project.readme]
file = "README.rst"
content-type = "text/x-rst; charset=UTF-8"
# Add here related links, for example:

[project.urls]
Homepage = "https://github.com/pyscaffold/pyscaffold/"
Documentation = "https://pyscaffold.org/"
# Source = https://github.com/pyscaffold/pyscaffold/
# Changelog = https://pyscaffold.org/en/latest/changelog.html
# Tracker = https://github.com/pyscaffold/pyscaffold/issues
# Conda-Forge = https://anaconda.org/conda-forge/pyscaffold
# Download = https://pypi.org/project/PyScaffold/#files
# Twitter = https://twitter.com/PyScaffold
# Change if running only on Windows, Mac or Linux (comma-separated)
# Add here all kinds of additional classifiers as defined under
# https://pypi.org/classifiers/

[project.optional-dependencies]
# Add here additional requirements for extra features, to install with:
testing = [
    "setuptools",
    "pytest",
    "pytest-cov",
    "pytest-compare[pandas]",
]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-compare[pandas]",
    "black",
    "pre-commit",
    "pylint",
]
docs = [
    "docutils",
    "sphinx",
    "sphinx-rtd-theme",
]

[project.entry-points]
# Add here console scripts like:
# console_scripts =
# script_name = imputegaps.module:function
# For example:
# And any other entry points, for example:
# pyscaffold.cli =
# awesome = pyscaffoldext.awesome.extension:AwesomeExtension

[project.scripts]
statstools = "imputegaps.main:run"

[tool.setuptools]
zip-safe = false
include-package-data = true
package-dir = {"" = "src"}
# Require a min/specific Python version (comma-separated conditions)
# python_requires = >=3.8
# Add here dependencies of your project (line-separated), e.g. requests>=2.2,<3.0.
# Version specifiers like >=2.2,<3.0 avoid problems due to API changes in
# new major versions. This works if the required packages follow Semantic Versioning.
# For more information, check out https://semver.org/.
platforms = ["any"]
license-files = ["LICENSE.txt"]

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests"]
namespaces = true

[tool.pytest.ini_options]
# Specify command line options as you would do when invoking pytest directly.
# e.g. --cov-report html (or xml) for html/xml output or --junitxml junit.xml
# in order to write a coverage file that can be read by Jenkins.
# CAUTION: --cov flags may prohibit setting breakpoints while debugging.
# Comment those flags to avoid this pytest issue.
addopts = """
--cov imputegaps --cov-report term-missing
--verbose"""
norecursedirs = [
    "dist",
    "build",
    ".tox",
]
testpaths = ["tests"]
# Use pytest markers to select/deselect specific tests
# markers =
# slow: mark tests as slow (deselect with '-m "not slow"')
# system: mark end-to-end system tests

[tool.devpi.upload]
# Options for the devpi: PyPI server and packaging tool
# VCS export must be deactivated since we are using setuptools-scm
no_vcs = "1"
formats = "bdist_wheel"

[tool.flake8]
# Some sane defaults for the code style checker flake8
max_line_length = "120"
extend_ignore = "E203, W503"
# ^  Black-compatible
# E203 and W503 have edge cases handled by black
exclude = """
.tox
build
dist
.eggs
docs/conf.py"""

[tool.pyscaffold]
# PyScaffold's parameters when the project was created.
# This will be used when updating. Do not change!
version = "4.5"
package = "imputegaps"
# This file is used to configure your project.
# Read more about the various options under:
# https://setuptools.pypa.io/en/latest/userguide/declarative_config.html
# https://setuptools.pypa.io/en/latest/references/keywords.html
